requirements.txt
faiss-cpu==1.7.4
huggingface-hub==0.33.0
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.25
openai==1.91.0
python-dotenv==1.0.0
python-magic==0.4.27
requests==2.32.4
sentence-transformers==4.1.0
streamlit==1.46.0
tiktoken==0.9.0
unstructured==0.9.2